Collision and Connected Passage Ways
Finished the Map generation, gave the ability to connect passageways.
What's left to do is indicate maybe different areas as certain spots or
where most likely to find certain things. Added Collisions and gave a
test scene where you can see for yourself the collisions occuring, if
the box spawns inside a wall just move it out of the wall. Use the arrow
keys to move it around.